Fixed second bug
In this commit, i fixed the second bug that didn't allow for users to reset the board after clicking the reset button or clicking on a new puzzle. I achieved this by  creating a "resetPuzzleBoard" function that moves all puzzle pieces back to the puzzle pieces container. This function should be called when the "Reset" button is clicked or when a new puzzle is chosen. I also included the reset functionality in the "changeBGImage" function ensuring that the puzzle board is reset every time a new puzzle is chosen by clicking on a thumbnail.